welcome limbo version.
this also applies a minor fix to the c version.  the bug caused
retries of password changes to fail.

note that the limbo version has the same structure as the c version.
this is meant to make it easier to audit them both, and apply fixed
from one to the other.